Move context-store instrumentation into AgentTransformerBuilder (reduce abstractions and unnecessary static collections)
